can i speak to your manager?
new entity manager to abstract entities. previously, an entity would require; a rawmodel, a modeltexture, a texturedmodel, and to top it off, a render call in the while loop. now you just need an entitymanager and a entitymanager render call and now you can have as many entities just with this. entities are created with a .ntt file stored in src/entities/ntt. next i want to make some documentation so i can have people use this engine with me without needing too much assistance